chore(lists): send email notification to users when a new submission is made
This commit adds a new feature to send an email notification to users when a new submission is made. The `sendNewSubmissionEmail` function is moved from `src/server/components/lists/controllers/ingest/helpers.ts` to `src/server/components/lists/helpers.ts` and is updated to accept a `listId` parameter instead of a `ListDataForSubmissionEmail` object. The function retrieves the list object from the database using the `listId` and sends an email notification to all users in the `jsonData.users` array. The `listsConfirmApplicationController` and `ingestPostController` functions are updated to call the `sendNewSubmissionEmail` function after a new submission is made.
